refactor dataRow components for transaction details
