+++ i18n/locales/en/common.json
@@ -293,6 +293,7 @@
   "No peers information": "No peers information",
   "No versions distribution information": "No versions distribution information",
   "No votes in queue.": "No votes in queue.",
+  "Nonce": "Nonce",
   "Not Empty": "Not Empty",
   "Not enough LSK to pay for the transaction.": "Not enough LSK to pay for the transaction.",
   "Not forging": "Not forging",
