+++ src/components/screens/transactionDetails/dataRows/accountInfo.test.js
