+++ src/components/screens/transactionDetails/dataRows/index.js
@@ -1,0 +1,15 @@
+/* eslint-disable import/prefer-default-export */
+
+export {
+  Illustration, Sender, Recipient, Fee, Amount, Date, TransactionId,
+} from './baseComponents';
+
+export {
+  SenderAndRecepient,
+  TransactionIdAndAmount,
+  RequiredSigsAndFee,
+  DateAndConfirmations,
+  MessageAndNonce,
+  AmmountAndDate,
+  FeeAndConfirmations,
+} from './dataRows';
